backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.1
alembic==1.12.1
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
fake-useragent==1.3.0
pandas==2.1.3
aiohttp==3.9.0
asyncio==3.4.3
tqdm==4.66.1
anthropic==0.7.0
stripe==7.8.1